--- notify/providers/onesignal/settings.py
@@ -1,6 +1,0 @@
-from navconfig import config
-
-# OneSignail
-ONESIGNAL_PLAYER_ID = config.get("ONESIGNAL_PLAYER_ID", fallback="")
-ONESIGNAL_OS_APP_ID = config.get("ONESIGNAL_OS_APP_ID", fallback="")
-ONESIGNAL_OS_API_KEY = config.get("ONESIGNAL_OS_API_KEY", fallback="")
