--- notify/providers/twilio/settings.py
@@ -1,6 +1,0 @@
-from navconfig import config
-
-# Twilio credentials
-TWILIO_ACCOUNT_SID = config.get("TWILIO_ACCOUNT_SID")
-TWILIO_AUTH_TOKEN = config.get("TWILIO_AUTH_TOKEN")
-TWILIO_PHONE = config.get("TWILIO_PHONE")
