--- notify/providers/twitter/settings.py
@@ -1,7 +1,0 @@
-from navconfig import config
-
-# Twitter Tweets:
-TWITTER_ACCESS_TOKEN = config.get("TWITTER_ACCESS_TOKEN")
-TWITTER_TOKEN_SECRET = config.get("TWITTER_TOKEN_SECRET")
-TWITTER_CONSUMER_KEY = config.get("TWITTER_CONSUMER_KEY")
-TWITTER_CONSUMER_SECRET = config.get("TWITTER_CONSUMER_SECRET")
